Move kickstart support to the FCoE module
The kickstart support for FCoE devices was moved to the FCoE module.
